Parameterize `run_once` to enable better startup
By passing both the command name and the arguments of the command
separately, we can properly grep for the process with pgrep and avoid
double starting processes like cbatticon or remind.

This does require a change to the autostart file, namely, all lines must
contain a single `;` regardless of whether the command is taking arguments
or not.