emacs: tramp: zsh: unset options when using TRAMP
TRAMP would never recognize the remote shell when connecting to a remote
system using `zsh`.  Following some troubleshooting tips from the
[wiki][0], this configuration can now appropriate connect to remote
systems, given the `zsh` configuration is deployed.

It might be of dubious value to continue to use `zsh` in this context.
A possibly better solution is to simply use `bash` when using TRAMP.

[0]: https://www.emacswiki.org/emacs/TrampMode

Signed-off-by: Kenny Ballou <kballou@devnulllabs.io>