emacs: org: modify and move capture templates
Instead of placing much of the various capture (meta)data into the org
entry itself, move these values into the `:PROPERTIES:` drawer of the
entry.  Some of this is borrowed from [John Wiegley's talk][0] on
org-mode, specifically, the `:CREATED:` lines and the RESPOND template.

Furthermore, given the general increase in complexity of the templates,
move the templates into files that are loaded by org-mode at capture.
[[1],[2],[3]].

[0]: https://karl-voit.at/2019/10/26/all-things-org/

[1]: https://emacsnyc.org/assets/documents/how-i-use-org-capture-and-stuff.pdf

[2]: https://joshrollinswrites.com/help-desk-head-desk/org-capture-in-files/

[3]: https://irreal.org/blog/?p=8161

Signed-off-by: Kenny Ballou <kballou@devnulllabs.io>