emacs: org: specifically set org-id-locations
This occasionally gets set a file in the nix store, which then makes
closing emacs impossible without resetting it.

To be figured out is how it gets set to a file in the nix store.  I
usually often only when attempting to kill the emacs daemon, which may
days or weeks after the variable has been updated.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>