emacs: disable icons in dired and ibuffer
Sometimes I want to open folders with _lots_ of files.  Motivation for
this removal was a folder with 5k files.  The minor benefit of seeing
some icons on occasion is not worth the 5 to 10 minutes emacs completely
locks up trying to display some icons.  Spamming ^G and sending SIGUSR2
is maybe the reason it's _only_ 5 to 10 minutes...

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>