systems: daeva: remove guix gc and nix gc jobs
Because the GC roots don't seem to be properly set, these jobs tend to
nuke project builds and caches for different environments.  At best, it
just means the projects need to be rebuilt.  Worse, it messes with the
caches and breaks the projects.

Since these are relatively stable, these jobs do not need to run daily.
On demand should be sufficient.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>