config: emacs: use flymake over flycheck
Flycheck has been causing me many problems as of late.  For some large
files, for example, flycheck must be disabled entirely because the
selected checkers choke the buffer, making it impossible to
interact/type into the buffer (thus far, this does not appear to be a
problem with flymake with the same checkers).  Othertimes, flycheck does
not (and cannot be manually forced to) recheck a buffer with selected
checkers (only way that seems to work for flycheck is to toggle the
mode).  However, flymake has the ability to be manually run.  As per
availability of checkers for different major modes, this seems to be
less of any issue since flymake-flycheck can generate flymake backends
from flycheck checkers.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>