config: guix: shell auth: install template unless file exists
Install a basic template for `shell-authorized-directories`. If a
"real" version exists (which will not be checked in anymore),
concatenate it into the template.

While this is not private in the sense that the computed file is still
in the store, it does serve a purpose of not having to commit the whole
file every time.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>