config/emacs: add accessibility to latex
LaTeX generated documents from Org-Mode now include the document
metadata which mostly enables tagged PDF's by default.  This massively
improves the accessibility of PDF's without a ton of work.  Still
missing is usually a decent treatment of images and links (often Canvas
will read a link or footnote of a link to be an image, which downgrades
the score).

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>