config/emacs: add handoutonly export processing
For org-teach, which overrides the export function to produce handouts
and slides from the same document, I occasionally produce handout only
lecture notes.  Since producing slides that will never be used slows the
flow, I wanted to add an option that allows me to override the behavior
of the custom export to read the tags of the file and produce the
correct results.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>