config/emacs: org: add function to visit attach files
From an org-attach directory, we want to find the associated file.
Since attach directories follow the UUID naming scheme where the first
byte is a parent folder and the remaining bytes are the next subfolder,
we combine the folder names into the original UUID and query the hash
table for org id locations.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>