packages/emacs: Prefer stable versions
When available, prefer to use stable versions of software.  We're on the
edge of package bankruptcy...  Many packages have been commented out or
removed.  Likely more to be removed as we cull unused packages.

This fixes an issue, among others, where the MELPA version of `magit`
links in with the stable version of `transient`, which [we should most
certainly not do][0].

[0]: https://github.com/magit/magit/discussions/5375