timers: remove guix gc "repair"
This is quite dangerous.  I have learned the hard way that this is not
something that should be run premeptable and peridoically...  I
suspect my hardware less now after reading the documentation on this
command.  XFS may still be viable, but I'm not likely to switch back
unless I end up re-building the machine.

Signed-off-by: Kenny Ballou <kb@devnulllabs.io>